refactor: convert unmute.js to ts
